--- cmd/provisioners/ftl-provisioner-cloudformation/main.go
@@ -1,35 +1,0 @@
-package main
-
-import (
-	"context"
-
-	"connectrpc.com/connect"
-
-	ftlv1 "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1"
-	"github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1beta1/provisioner/provisionerconnect"
-	"github.com/TBD54566975/ftl/common/plugin"
-)
-
-type CloudformationProvisioner struct{}
-
-var _ provisionerconnect.ProvisionerPluginServiceHandler = (*CloudformationProvisioner)(nil)
-
-func NewCloudformationProvisioner(ctx context.Context, config struct{}) (context.Context, *CloudformationProvisioner, error) {
-	return ctx, &CloudformationProvisioner{}, nil
-}
-
-func (c *CloudformationProvisioner) Ping(context.Context, *connect.Request[ftlv1.PingRequest]) (*connect.Response[ftlv1.PingResponse], error) {
-	return &connect.Response[ftlv1.PingResponse]{}, nil
-}
-
-func main() {
-	plugin.Start(
-		context.Background(),
-		"ftl-provisioner-cloudformation",
-		NewCloudformationProvisioner,
-		"",
-		provisionerconnect.NewProvisionerPluginServiceHandler,
-	)
-}
-
-func ptr[T any](s T) *T { return &s }
